Lecture 10 Emoji Art
Create an iPad application from scratch: Emoji Art.  This is an opportunity to review MVVM since this application has an entirely different architecture from Memorize.  It uses SwiftUI's drag and drop mechanism to set the Emoji Art document's background and to add emojis to the works of art you can create with it.