Lecture 14 Multithreading, Error Handling
Mobile application UIs need to always be responsive even though they may do things that can take a (relatively) long time (like downloading something from the internet). The Swift language has an integrated mechanism for making it clear which things need to happen "in the background" so as not to slow down the UI.  Handling unexpected errors is also important while doing this, so both topics are presented together in this lecture.